# Port naive LLL reduction checks to gr_mat; speed up fmpz_mat versions