The LLL code ends up calling ``fmpz_mat_is_reduced`` or ``fmpz_mat_is_reduced_with_removal`` as a fallback if the floating-point verifications fail. These functions implement the naive algorithm using rational arithmetic, so they can be extremely slow if the dimension is large. The LLL ought to be improved so that these fallbacks can be avoided (almost) entirely, but a slight improvement is to make them less horribly slow.

We thus port the naive algorithm to generics and make ``fmpz_mat_is_reduced`` / ``fmpz_mat_is_reduced_with_removal`` run some rounds with ``arb`` arithmetic before ultimately falling back on ``fmpq``.

Improvements on the ``fmpz_poly_factor`` benchmark:

Old:
```
H1 has 28 factors: 1669 ms
S7 has 1 factors: 139 ms
C1 has 32 factors: 580 ms
H2 has 6 factors: 230955 ms
S9 has 1 factors: 197224 ms
```

New:
```
H1 has 28 factors: 1585 ms
S7 has 1 factors: 110 ms
C1 has 32 factors: 520 ms
H2 has 6 factors: 218477 ms
S9 has 1 factors: 193651 ms
```